Remove CDChunkList
CDChunkList was a version of CDList that could be allocated with the
ContextMemoryAllocator. None of the CDChunkList instances was using this
feature, so this commit changes them to CDLists. Additionally, the use
of CDChunkList was problematic: if a CDChunkList was created at a
non-zero context level and the context was popped below that level,
subsequent pushes would corrupt the context memory. This happened with
resetAssertions(), which pops down to level 0 while some CDChunkLists
were created at level 1. See also the added test case
"testPopBelowLevelCreated" for CDLists for illustration.