Rewrites for substr of strings of length one
This commit adds a rewrite for substrings of strings of length one to
the empty string if it can be shown that it is not possible that the
start position and the length are both greater than zero.  The commit
introduces a function `checkEqAndEntailArithUnsat` that can be used to
do the required reasoning.