Better normalization of string concatenation
This commit adds a way to normalize string concatenations such as:

```
(str.++ ... (str.replace "A" x "") "A" (str.substr "A" 0 z) ...) --> (str.++ ... [sort those 3 arguments] ... )
```

This is based on the observation that both `(str.replace "A" x "")` and
`(str.substr "A" 0 z)` either return the empty string or "A", so between
those terms and the string constant "A" the order doesn't matter.

To implement this normalization, this commit adds the function
`getStringOrEmpty()`, which overapproximates the possible values of a
term, assuming that the possible values are some value x and the empty
string.