Warn about trailing spaces in src/Makefile.am
The contrib/new-theory script expects no trailing whitespaces in
src/Makefile.am otherwise it inserts the list files of the new theory at
a wrong place in the file. This commit adds a check for that in the
beginning of the script. Since contrib/new-theory is run as part of our
Travis tests, this will allow us to detect trailing spaces and fix them
rather than silently ignoring them.