Fix get-unsat-core detection in regression script
Due to a typo in the regression script, the (get-unsat-core) command was
not recognized (the script was looking for (get-unsat-cores)), so it
tried to run a regression script containing (get-unsat-core) even when
CVC4 was compiled without proof support. This commit fixes the typo and
additionally changes the script to look for the pattern in the benchmark
file instead of the metadata file (i.e. the .expect file if it exists or
the benchmark file otherwise). The second fix is not strictly necessary
for our current regressions but might have lead to surprising outcomes
in the future.