RemoveTermFormulas: Remove ContainsTermITEVisitor
ContainsTermITEVisitor was not really used by RemoveTermFormulas nor
directly by anyone using RemoveTermFormulas (at least not since
1f4b954a2cc7667a56a3007fa75c125fba93ed23 commented out the code calling
containsTermITE from within RemoveTermFormulas). The only purpose was to
pass it to an ITEUtilities in TheoryEngine. This commit moves the
ownership to ITEUtilities and creates it directly wihin ITEUtilities. As
far as I can tell, there is only one place where ITEUtilities and
RemoveTermFormulas are created (and there were no errors complaining
about mismatching constructor arguments), so this move should be safe
and simplifies the code a bit (simpler depedencies, simpler
initialization of ITEUtilities, less code).