Remove (dummy) SMT1 printer
CVC4 currently claims that it supports SMT 1.x as an output language but
the implementation is just calling the SMT 2.5 printer. This commit
removes that dummy implementation. It also makes CVC4 output an error
when the user selects SMT 1.x as an output language and removes that
option from the help text.