Track input language in a single place
Currently, we store and update the input language in two places: the
Input object and the options. The language stored in the Input object
was only really used for SMT2 inputs. Storing the input language in two
places means that we have to update the language in two places, which
duplicates code (e.g. [0] and [1]). This removes the duplicate code and
changes the code to track the input language in the options only. This
commit is preparation for fixing #1969.

[0] https://github.com/CVC4/CVC4/blob/74c1ad7e4a8e93316b7555ac8a1b88ee777335e2/src/smt/smt_engine.cpp#L2359-L2381
[1] https://github.com/CVC4/CVC4/blob/74c1ad7e4a8e93316b7555ac8a1b88ee777335e2/src/parser/smt2/Smt2.g#L1131-L1146