Remove outdated references to TLS
Commit 546d795470ca7c30fc62fe9b6c7b8e5838e1eed4 removed configure-time
checks for compiler support for thread local storage because C++11
supports it by default. The commit did not quite remove all traces (note
that the `USING_TLS` macro was already broken at the time of that commit
because the configure script was setting `CVC4_TLS` instead of `TLS`).
This commit removes the remaining traces and updates the AUTHORS file to
just refer to the individual *.m4 files for authorship since some of
them have a large number of authors.